# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Teensy
  description: Tiny Url Generator
  version: 1.0.0
  contact:
    name: Tejas Ozarkar
    email: tejozarkar@gmail.com
servers:
  - url: http://localhost:8080
    description: Local Environment
  - url: http://teensyapi-env.eba-k2xfeapv.ap-south-1.elasticbeanstalk.com
    description: Production Environment
paths:
...
  /url/{shortUrl}:
    get:
      summary: Get Long Url
      operationId: getLongUrl
      responses:
        200:
          description: Return Long Url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponseDto'
      parameters:
        - name: shortUrl
          in: path
          description: Short Url
          required: true
          schema:
            type: string
  /url:
    post:
      summary: Crate Tiny Url
      operationId: createTinyUrl
      requestBody:
        description: Create Tiny Url
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlRequestDto'
        required: true
      responses:
        200:
          description: Url Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponseDto'
...
components:
  schemas:
...
    UrlRequestDto:
      type: object
      properties:
        longurl:
          type: string
        expirydate:
          type: string
    UrlResponseDto:
      type: object
      properties:
        id:
          type: integer
        shorturl:
          type: string
        longurl:
          type: string
        expirydate:
          type: string
...